Logout functionality in place
