+++ services/ui-src/src/components/Header.tsx
@@ -1,7 +1,5 @@
 import React, { RefObject, useState, useEffect, useRef } from "react";
 import { Link, useHistory, useLocation } from "react-router-dom";
-import { Auth } from "aws-amplify";
-import { AwsCognitoOAuthOpts } from "@aws-amplify/auth/lib-esm/types/Auth";
 import { Button, SkipNav } from "@cmsgov/design-system";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
@@ -16,6 +14,7 @@ import {
   getUserRoleObj,
   inFlightRoleRequestForUser,
 } from "cmscommonlib";
+import { getRegisterUrl, getSignInUrl, logout } from "../libs/logoutLib";
 import { getCurrentRoute } from "../utils/routeUtils";
 import config from "../utils/config";
 import { Alert, UsaBanner } from "@cmsgov/design-system";
@@ -26,56 +25,6 @@ import { ONEMAC_ROUTES, ROUTES as RouteList } from "cmscommonlib";
 import HamburgerMenu from "../components/HamburgerMenu.js";
 
 import { animated, useSpring } from "@react-spring/web";
-
-/**
- * Get the sign in URL used with OKTA.
- * @returns the signin URL
- */
-export function getSignInUrl() {
-  const authConfig = Auth.configure();
-  const { domain, redirectSignIn, responseType } =
-    authConfig.oauth as AwsCognitoOAuthOpts;
-  const clientId = authConfig.userPoolWebClientId;
-  const url = `https://${domain}/oauth2/authorize?identity_provider=Okta&redirect_uri=${redirectSignIn}&response_type=${responseType}&client_id=${clientId}`;
-  return url;
-}
-
-/**
- * Get the register URL depending on the current domain.
- * @returns the register URL
- */
-function getRegisterUrl() {
-  const currentDomain = window.location.hostname;
-  let registerUrl = "https://test.home.idm.cms.gov/";
-
-  // TODO remove the 'spa.cms.gov' and 'spa-val.cms.gov' as options
-  // after the rebrand has changed the domain to onemac
-  if (currentDomain === "onemac.cms.gov" || currentDomain === "spa.cms.gov") {
-    registerUrl = "https://home.idm.cms.gov/";
-  } else if (
-    currentDomain === "onemacval.cms.gov" ||
-    currentDomain === "spa-val.cms.gov"
-  ) {
-    registerUrl = "https://impl.home.idm.cms.gov/";
-  }
-
-  return registerUrl;
-}
-
-/**
- * Logout the user.
- */
-function logout(isLoggedInAsDeveloper?: boolean) {
-  const authConfig = Auth.configure();
-  Auth.signOut();
-  if (isLoggedInAsDeveloper) {
-    window.location.replace(
-      (authConfig.oauth as AwsCognitoOAuthOpts).redirectSignOut
-    );
-  } else {
-    window.location.href = getRegisterUrl();
-  }
-}
 
 /**
  * Hook that alerts clicks outside of the passed ref
